refactor(antithesis): move payload helpers into module
